requirements.txt
# Simple PDF to JSON Bot
groq>=0.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image

## old code

#pdfplumber
#python-dotenv
#groq
#pytesseract
#pdf2image
#Pillow